The JuliaDispatch.Analysis Module

This module contains a few functions for basic analysis of simulations, including a finite difference gradient method, similar to numpy.gradient.

JuliaDispatch.Analysis.average — Method

average(snap::Dict; iv::Union{Int, String})

Average of quantity iv over all patches

JuliaDispatch.Analysis.diff_quotient — Method

diff_quotient(F::AbstractArray; spacing::Union{AbstractArray, Number}=1, dim=1)

Compute the difference quotient of F along dimension dim using finite difference with spacing giving the coordinate distance between elements in F. spacing can be a number, which is equivalent to equal spacing among all points, or it can be an array with spacing between the values. Must then be equal to size(F, dim).

JuliaDispatch.Analysis.gradient — Method

gradient(F::AbstractArray; spacing::Union{AbstractArray, Number}=1)

Compute the difference quotient of F along all dimensions. spacing can be a number specifying uniform sample distance for all dimensions, or an array or, a collection of ndims(F) arrays for each value in each dimension.

JuliaDispatch.Analysis.plaverage — Method

plaverage(snap; iv=0, dir=3, all=false, nslices=nothing)

Compute and return the average value of quantity iv in each plane slice perpendicular to direction dir.

JuliaDispatch.Analysis.time_evolution_average — Method

time_evolution_average(;iv = 0, run="", data="../data",
                        all=false, verbose=0, tspan=nothing)

Time evolution of quantity iv averaged over entire computational domain.

JuliaDispatch.Analysis.time_evolution_plaverage — Method

time_evolution_plaverage(;iv = 0, run="", data="../data", dir=3, tspan=nothing,
           all=false, verbose=0, nslices=nothing)

Time evolution of plane average of quantity iv along direction dir. Returns a Vector containing the times of each snapshot, a Vector containing the axis coordinates along direction dir, and a Matrix with the averages at each coordinate (axis 1) at each time (axis 2).