The JuliaDispatch.Buffers Module

The purpose of this module is to stitch together all the patch data of a given quantity in to a single array. The module contains functionality for doing this for both a unigrid and a mesh-refined simulation, and for both a 3D and 2D domain. Additionally, it has the ability to down- or upscale the stitched data using gridded bi/tri-linear interpolation, which allows for faster and more memory efficient data extraction.

As with the other modules, after having imported ´JuliaDispatch` into your workspace, do

julia> using JuliaDispatch.Buffers

unigrid_plane

The unigrid_plane function takes in a snapshot and a single physical coordinate at some x/y/z-value and attempts to combine the data using the patches that are found within the given plane. This function returns the entire data set, meaning it does not perform any interpolation when stitching together the data apart from interpolating patch data between planes to get the values at the given coordinate.

JuliaDispatch.Buffers.unigrid_plane — Function

unigrid_plane(snap::Dict; iv::Union{Int, String}, x::Float, y::Float, z::Float)

Return a 2d array of joined patch data in a slice at x/y/z. All patches in given plane must have same shape, size, and number of cells. Returns a Dict if iv is a collection, otherwise a Matrix.

unigrid_volume

Similar to unigrid_plane except that it returns a 3D array with the entire simulation data set for the given quantity.

JuliaDispatch.Buffers.unigrid_volume — Function

unigrid_volume(snap::Dict; iv::Union{Int, String}=0, all::Bool=false, verbose::Int=0)

Return a 3D array of joined patch data of quantity . All patches must have same shape, size, and number of cells.

#Arguments:

snap::Dict, snapshot

#Kwargs: - iv::Union{String, Int, Array}, what quantity(/ies) to buffer. if array, output will be a dictionairy with keys equal to the input variable names, and values as the 2d-buffers. - all::Bool, whether to include guard zones (if available)

amr_plane

If the simulation uses mesh-refinement or you want to up/downscale the data from a unigrid experiment, use the amr_plane function. In addition to the standard parameters of unigrid_plane , this function also accepts the dimensions of the returned array containing the interpolated data. The dimensions are given with the keyword argument dims and can be an Integer denoting a square array, or a length-2 array or tuple denoting the dimensions in axis 1 and axis 2.

JuliaDispatch.Buffers.amr_plane — Function

amr_plane(snap::Dict; iv::Union{Int, String}, x::Float, y::Float, z::Float,
         Log::Bool, dims::Union{Int, Tuple}, verbose=0)

Return a 2D array containing interpolated data of quantity iv in a slice x/y/z from all patches in a given snapshot. If dims is an Int the resulting array will be of size (dims, dims). If dims is a length-2 array, the array will have size (dims...).

amr_volume

Similar to amr_volume but for 3D data.

JuliaDispatch.Buffers.amr_volume — Function

amr_volume(snap::Dict; iv::Union{Int, AbstractArray, String}, all::Bool,
        dims::Union{Tuple, Int}, verbose::int)

Return a 3D array containing interpolated data of quantity iv from all patches in a given snapshot. If dims is an Int the resulting array will be of size (dims, dims, dims). If dims is a length-3 array, the array will have size (dims...).