The JuliaDispatch.Dispatch Module

This module contains the snapshot-function, which is used for reading a snapshot and parsing the patches.

After having imported the JuliaDispatch module into your workspace, simply do

julia> using JuliaDispatch.Dispatch

The function snapshot is then available. See end of page for documentation and examples.

When reading in a snapshot with the snapshot function, a dictionary will be returned. All the properties of the snapshot can be accessed using square brackets and a string with the name of the property. For example, if you want to know the time at which the snapshot was taken, do

println(snap["time"])

All the patches and their metadata are stored in the "patches" key of the snapshot. This returns a vector with each entry containing the patch metadata in the form of another dictionary. For example

patch1 = snap["patches"][1]
println(patch1["id"])

extracts the first patch and prints its id.

JuliaDispatch.Dispatch.snapshot — Function

snapshot(iotu::Int; run::String, data::String, verbose::Int)

Parses patches and returns a Dict with all properties of snapshot iout.

#Arguments:

iout::Int, snapshot ID

#Kwargs:

data::String, path to data (snapshots), default "..\data"

run::String, path to snapshots folders relative to data, default ""