The JuliaDispatch.Graphics Module

This module contains various functions for visualizations, including 1D, 2D and 3D plots and simple animations.

To begin, first import the module into your workspace

julia> using JuliaDispatch.Graphics

Assuming you already have a snapshot loaded, you can then call the various plotting functions to visualize quantities. All the plotting functions accepts keyword arguments which are supported by the Plots.jl GR backend. See [documentation][http://docs.juliaplots.org/latest/generated/gr/] for all possible keyword arguments. This means that you can set title, axis labels, fontsizes, linestyle (lineplot, scatter, heatmap, surface, etc) manually.

sliceplot

The sliceplot function is the main function for plotting a quantity in a given plane. The function is very flexible and includes, among other things, the possibility to zoom in on certain regions by using the center and/or width keyword arguments. If the given combination of center and width extends beyond a boundary, the function will look for periodicity in the snapshot and stitch together the data beyond the boundary.

By default, sliceplot assumes the snapshot data is not mesh-refined. If it is mesh-refined, make sure the unigrid keyword argument is set to false.

It it possible to downscale the data to save memory by setting the resample keyword argument to true. This will rescale the data to a size defined by the dims keyword argument, which is by default set to 300. dims can be either an integer, which rescales the data to have the same number of grid points in each dimension, or a 2D collection (Tuple or Array) with the number of grid points for each axis.

JuliaDispatch.Graphics.sliceplot

plotvaluesalong

To plot values of a quantity along a given axis through a given point, use the function plot_values_along. This function takes in a snapshot, a point (in the form of a length-3 array) as arguments, and the quantity iv and the axis direction dir as keyword arguments, in addition to any keyword argument accepted by the Plots.plot() function.

JuliaDispatch.Graphics.plot_values_along

histogram_along

The function histogram_along computes and plots a histogram of a quantity iv along a direction dir through a point pt. By default the histogram shows the absolute frequency (counts), but can be changed to a probability density or probability by setting the keyword argument norm to, respectively, :pdf or :probability. Additionally you define the bins with the bins keyword argument.

JuliaDispatch.Graphics.histogram_along

volume

work in progress

anim_plane

To produce a time-evolution animation of given quantity in a 2D plane, use the anim_plane function. This function accepts any keyword arguments accepted by sliceplot, in addition to the tspan keyword argument which can be a tuple of numbers denoting at which time (which snapshot) to start and end the simulation. Additionally, the step keyword arguments defines whether to use all snapshots or to skip with a given step value.

JuliaDispatch.Graphics.anim_plane — Function

anim_plane(;data="../data", run="", x = nothing, y = nothing, z = nothing, iv=0, 
            tspan=nothing, unigrid=true, step=1, savepath=nothing, verbose = 0, kw...)

Animate a plane of quantity iv at a slice x/y/z. If tspan tuple is not given, all snapshots will be used, otherwise only snapshots with time within the given tspan will be loaded. A frame will be recorded every step, and the resulting animation will be saved to savepath.

anim_pane

If you want animate a pane through the computational domain in a certain direction, call the anim_pane function. This function accepts any keyword argument supported by sliceplot in addition to the nframes keyword argument which denotes how many frames will be written to the output video file. By default a pan will be made from the minimum value of the given axis to the maximum value, but this can be overwritten by setting the keyword arguments start and/or stop to desired values. Additionally you can set reverse to true if you would like to pan from the maximum value to the minimum.

JuliaDispatch.Graphics.anim_pane — Function

anim_pane(snap; ax=1, nframes=10, unigrid=true, iv=iv, reverse=false
          start=nothing, stop=nothing, verbose=0, savepath = nothing, kw...)

Plot and animate a pane through axis ax from start to stop with nframes frames of quantity iv. Output is saved to savepath. Will go from low to high axis values unless reverse is set to true.

#Examples

julia> anim_pane(snap, ax=3, iv="d", nframes=100, start=0.0, stop=-10.0, savepath="test.gif", reverse=true)