The JuliaDispatch.Select Module

This module contains functionality for locating and extracting patches and data at certain locations within the domain.

JuliaDispatch.Select.box — Method

box(patch::Dict; iv::Union{String, Int}, all::Bool, verbose::Int)

Return 3D-array of data of quantity iv from patch. Shorthand for patch["var"](iv, verbose=verbose, all=all).

Examples

julia> density_box = box(patch, iv="d")

JuliaDispatch.Select.corner_indices — Method

corner_indices(snap, patch; dir=1)

Get the corner indices of a given patch in a given direction.

JuliaDispatch.Select.corner_indices_all — Method

corner_indices_all(snap, patch)

Get the corner indices of a given patch.

JuliaDispatch.Select.indices_and_weights — Function

indices_and_weights(p::Dict, point = [0.5, 0.5, 0.5], iv = 0)

Return indices and interpolation weights for a point in a patch p.

JuliaDispatch.Select.is_inside — Method

is_inside(p::Dict, point::Array)

Return true/false depending on if point is inside patch.

JuliaDispatch.Select.patch_at — Function

patch_at(pp, point=[0.5, 0.5, 0.5]; verbose=0)

Find the patch that contains a given point.

JuliaDispatch.Select.patches_along — Function

patches_along(snap, point=[0.5, 0.5, 0.5]; dir = 1, verbose=0)

Get the patches along a given direction through a point

JuliaDispatch.Select.patches_in — Method

patches_in(snap::Dict, span::Tuple)

Return a Vector of patches that lie within a domain spanned by span.

JuliaDispatch.Select.patches_in — Method

patches_in(snap::Dict; x::Number=nothing, y::Number=nothing, z::Number=nothing)

Return a Vector of patches that lie within a given plane.

JuliaDispatch.Select.plane — Method

plane(patch::Dict; iv::Union{Int, String}, x::Float, y::Float, z::Float,
            Log::Bool, all::Bool)

Find the plane values of a patch at a given slice x/y/z by interpolating neighbouring planes.

Examples

julia> density_plane = plane(patch, iv="d", z=0.0)

JuliaDispatch.Select.values_along — Function

values_along(snap::Dict, point::Array=[0.5, 0.5, 0.5]; dir::Int=1, iv::Union{String, Int}=0, var=nothing,
                verbose::Int = 0, all::Bool = false)

Return s, f(s) with s the coordinates and f the values in the iv slot of data, taken along the direction v – so far restricted to axis values

JuliaDispatch.Select.values_along2 — Function

values_along2(snap::Dict, point::Array=[0.5, 0.5, 0.5]; dir::Int=1, iv::Union{String, Int}=0, var=nothing,
                verbose::Int = 0, all::Bool = false)

Return s, f(s) with s the coordinates and f the values in the iv slot of data, taken along the direction v – so far restricted to axis values

JuliaDispatch.Select.values_in — Function

values_in(snap::Dict, point::Array=[0.5, 0.5, 0.5]; dir::Int=1, iv::Union{String, Int}=0, var=nothing,
verbose::Int = 0, all::Bool = false)

Return s, f(s) with s the coordinates and f the values in the iv slot of data, taken along the direction v – so far restricted to axis values