The JuliaDispatch.Utils Module

Utils contains smaller functions used by other functions within the JuliaDispatch environment.

JuliaDispatch.Utils._kw_extract — Method

_kw_extract(kw, dict)

If key from dict occurs in kw, pop them

JuliaDispatch.Utils.get_n_patches — Method

get_n_patches(iout; data="data", run="")

Return the number of patches in the given snapshot iout in the data/run-directory.

JuliaDispatch.Utils.get_n_snapshots — Method

get_n_snapshots(;run="", data="../data")

Returns the total number of snapshots in the given data folder

JuliaDispatch.Utils.get_new_snapshot — Method

get_new_snapshot(current_snap; data="data/", run="", sleeptime=10, maxsleep=100)

Return snapshot id with ID = current_snap + 1. Will wait for new snapshots as they are being produced. If program waits for more than maxsleep seconds without finding a new snapshot, the program will exit.

JuliaDispatch.Utils.get_snapshot_ids — Method

get_snapshot_ids(tspan::Tuple; data="../data", run="")

Return an array of IDs of all snapshots in the given data/run-folder whose time ∈ tspan

JuliaDispatch.Utils.get_snapshot_ids — Method

get_snapshot_ids(;data = "../data", run = "")

Return an array of IDs of all snapshots in the given data/run-folder.

JuliaDispatch.Utils.get_snapshot_time — Function

get_snapshot_time(iout=nothing; data="../data", run="")

Return the time of snapshot iout in the given data/run-folder or all snapshots if isnothing(iout).

JuliaDispatch.Utils.get_snapshot_times — Method

get_snapshot_times(; data="../data", run="")

Return the times of all snapshots in the given data/run-folder.