The JuliaDispatch.Dispatch Module

This module contains the snapshot-function, which is used for reading a snapshot and parsing the patches.

After having imported the JuliaDispatch module into your workspace, simply do

julia> using JuliaDispatch.Dispatch

The function snapshot is then available. See end of page for documentation and examples.

When reading in a snapshot with the snapshot function, a dictionary will be returned. All the properties of the snapshot can be accessed using square brackets and a string with the name of the property. For example, if you want to know the time at which the snapshot was taken, do

println(snap["time"])

All the patches and their metadata are stored in the "patches" key of the snapshot. This returns a vector with each entry containing the patch metadata in the form of another dictionary. For example

patch1 = snap["patches"][1]
println(patch1["id"])

extracts the first patch and prints its id.

The snapshot function will by default display a progressbar visualizing the progress in parsing the patches. To disable this, set the keyword argument progress to false. To suppress all printed information, which might be useful for when reading in a large number of snapshots at once, set verbose to -1.

JuliaDispatch.Dispatch.snapshot — Function

snapshot(iout::Int; run::String, data::String, progress::Bool=true, suppress::Bool=false,
         verbose::Int, copy::Bool=false, memmap::Int=1)

Parses patches and return a Dict with all properties of snapshot iout in the given data/dir-folder. Will display patch parsing progress unless progress is set to false.

JuliaDispatch.Dispatch.cache_snapshots_live — Function

cache_snapshots_live(;data="data/", run="", current_snap=0, sleeptime=10)

Read snapshots live as they are being produced by a simulation in the given data/run directory, caching the namelists for faster loading of snapshots later. Will start from snapshot current_snap. Will wait for sleeptime seconds to look again if no new snapshot is found. If program waits for more than maxsleep seconds without finding a new snapshot, the program will exit.

JuliaDispatch.Dispatch.get_snapshots — Function

get_snapshots(;data=data, run=run, tspan=nothing)

Returns a vector of parsed snapshots in the given data/run folder. If tspan is given, only snapshots with tspan[1] <= "time" <= tspan[2] will be parsed and returned.